--- public/index.d.ts
@@ -1,1 +1,0 @@
-export declare function getArkoseToken(): Promise<any>;
